buffer queries
