+++ docs/methods/query.md
@@ -2,11 +2,12 @@
 `query` can execute a query and serialize the results to a model.
 
 ## Parameters
-| name  | type        | description                                                                                   | optional     | default |
-|-------|-------------|-----------------------------------------------------------------------------------------------|--------------|---------|
-| sql   | `str`       | the sql query str to execute                                                                  | :thumbsdown: |         |
-| param | `ParamType` | params to substitute in the query                                                             | :thumbsup:   | `None`  |
- | model | `Any`       | the callable to serialize the model;  callable must be able to accept column names as kwargs. | :thumbsup:   | `dict`  |
+| name     | type        | description                                                                                   | optional     | default |
+|----------|-------------|-----------------------------------------------------------------------------------------------|--------------|---------|
+| sql      | `str`       | the sql query str to execute                                                                  | :thumbsdown: |         |
+| param    | `ParamType` | params to substitute in the query                                                             | :thumbsup:   | `None`  |
+ | model    | `Any`       | the callable to serialize the model;  callable must be able to accept column names as kwargs. | :thumbsup:   | `dict`  |
+ | buffered | `bool`      | whether to buffer reading the results of the query                                            | :thumbsup:   | `True`  |
 
 ## Example - Serialize to a dataclass
 The raw sql query can be executed using the `query` method and map the results to a list of dataclasses.
@@ -19,5 +20,16 @@ The raw sql query can be executed using the `query` method and map the results t
 You can get creative with what you pass in to the model kwarg of `query`
 ```python
 {!docs/../docs_src/methods/query/one_to_one_query.py!}
+```
+(This script is complete, it should run "as is")
+
+
+### Example - Buffering queries
+By default, `query` fetches all results and stores them in a lift (buffered).  By setting `buffered=False`, you can
+instead have `query` act as a generator function, fetching one record from the result set at a time.  This may be useful
+if querying a large amount of data that would not fit into memory, but note that this keeps both the connection and
+cursor open while you're retrieving results.
+```python
+{!docs/../docs_src/methods/query/query_unbuffered.py!}
 ```
 (This script is complete, it should run "as is")
