+++ docs_src/methods/query/query_unbuffered.py
@@ -1,0 +1,6 @@
+from pydapper import connect
+
+with connect("postgresql://pydapper:pydapper@localhost/pydapper") as commands:
+    data = commands.query("select * from task", buffered=False)
+    print(type(data))
+    # <class 'generator'>
