+++ pyproject.toml
@@ -36,6 +36,7 @@ pymssql = { version = "^2.2.3", optional = true }
 types-psycopg2 = { version = "^2.9.4", optional = true }
 types-pymssql = { version = "^2.1.0", optional = true }
 cached-property = "^1.5.2"
+pytest-lazy-fixture = "^0.6.3"
 
 [tool.poetry.dev-dependencies]
 black = "^21.12b0"
